propsを利用することなく異なる階層のコンポーネントとデータの共有を行えるようにした
