+++ src/context/ReportContext.tsx
@@ -1,0 +1,7 @@
+import React, { createContext } from "react";
+import { Report } from "../types";
+
+export const ReportContext = createContext<{
+    report: Report | null;
+    setReport: (report: Report) => void;
+} | null>(null);
