+++ src/pages/Evaluation.tsx
@@ -1,18 +1,21 @@
-import { useLocation } from "react-router-dom"
-import { Student } from "../types"
+import { useContext } from "react";
+import { Student } from "../types";
+import { ReportContext } from "../context/ReportContext";
 
 // ページ名称については検討の余地あり
 const Evaluation = () => {
-    const location = useLocation();
+
+    // ここで、ReportContextから取得した値を使う
+    const { report } = useContext(ReportContext);
 
     return (
         <div>
             <h1>Evaluation Page</h1>
             <p>評価対象のフォルダがインポートされたら、遷移されてくるページ</p>
 
-            <h2>コース名: {location.state.course}</h2>
+            <h2>コース名: {report.course}</h2>
             <h3>学生:</h3>
-            {location.state.students.map((student: Student) => {
+            {report.students.map((student: Student) => {
                 return (
                     <div key={student.id}>
                         <h4>{student.userId}</h4>
